Decouple core tests from image_converter dependencies
